[6920] Position error message above the filename, add configuration for getResponseError in SingleFileUpload's XHRUpload
